fix: trocando INTEGER para DECIMAL em 'price' e 'total' price do DB

+++ back-end/src/database/migrations/20230329011657-products.js
@@ -17,7 +17,7 @@ module.exports = {
 
       price: {
         allowNull: false,
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(4,2),
       },
 
       urlImage: {
